requirements.txt
oyaml>=1.0
pyyaml>=6.0
packaging~=23.2
icecream~=2.1.3